# go-web/web-architecture/cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.NameRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  handler.Request:
    properties:
      code:
        type: string
      color:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      published:
        type: boolean
      stock:
        type: integer
    required:
    - code
    - color
    - name
    - price
    - published
    - stock
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support Digital House
    url: http://www.swagger.io/support
  description: This is a simple API to handle products
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: Bootcamp Go Wave 6 - API
  version: "1.0"
paths:
  /products:
    get:
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized + invalid token
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: 'Not found '
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.Response'
      summary: Show the list of products
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Product to Store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store a product in db
      tags:
      - Products
  /products/{id}:
    delete:
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete product
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: This endpoint update the product name
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Product Id
        in: path
        name: id
        required: true
        type: integer
      - description: Product name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/handler.NameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update a product name
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Id product
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Product to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update a product
      tags:
      - Products
swagger: "2.0"